test: separate gguf tests to model dependent and model independent tests
