+++ test/modelDependent/functionary/gguf.test.ts
@@ -1,18 +1,15 @@
-import {describe, expect, it} from "vitest";
-import GGUFReadStream from "../../src/gguf/ggufParser/stream/GGUFReadStream.js";
-import GGUFParser from "../../src/gguf/ggufParser/GGUFParser.js";
-import GGUFFetchStream from "../../src/gguf/ggufParser/stream/GGUFFetchStream.js";
-import {getModelFile} from "../utils/modelFiles.js";
-import GGUFInsights from "../../src/gguf/GGUFInsights.js";
-import {getTestLlama} from "../utils/getTestLlama.js";
-import GGUFMetadata from "../../src/gguf/GGUFMetadata.js";
-
-const remoteGGUFModel = "https://huggingface.co/TheBloke/Falcon-180B-Chat-GGUF/resolve/main/falcon-180b-chat.Q6_K.gguf-split-a?download=true";
+import {describe, expect, it, test} from "vitest";
+import GGUFReadStream from "../../../src/gguf/ggufParser/stream/GGUFReadStream.js";
+import GGUFParser from "../../../src/gguf/ggufParser/GGUFParser.js";
+import {getModelFile} from "../../utils/modelFiles.js";
+import GGUFInsights from "../../../src/gguf/GGUFInsights.js";
+import {getTestLlama} from "../../utils/getTestLlama.js";
+import GGUFMetadata from "../../../src/gguf/GGUFMetadata.js";
 
 describe("GGUF Parser", async () => {
     const modelPath = await getModelFile("functionary-small-v2.2.q4_0.gguf");
 
-    it("Magic should be GGUF local model", async () => {
+    test("Magic should be GGUF local model", async () => {
         const stream = new GGUFReadStream(modelPath);
         const magic = await stream.readNBytes(4);
         const magicText = String.fromCharCode(...magic);
@@ -28,24 +25,6 @@ describe("GGUF Parser", async () => {
 
         expect(metadata).toMatchSnapshot();
     });
-
-    it("Magic should be GGUF remote model", async () => {
-        const stream = new GGUFFetchStream(remoteGGUFModel);
-
-        const magic = await stream.readNBytes(4);
-        const magicText = String.fromCharCode(...magic);
-
-        expect(magicText).toBe("GGUF");
-    });
-
-    it("should parse remote gguf model", async () => {
-        const stream = new GGUFFetchStream(remoteGGUFModel);
-
-        const ggufParser = new GGUFParser(stream);
-        const metadata = await ggufParser.parseMetadata();
-
-        expect(metadata).toMatchSnapshot();
-    }, {timeout: 0});
 
     it("should calculate GGUF VRAM Usage", async () => {
         const stream = new GGUFReadStream(modelPath);
